Refactor: Enhance Order Details and Improve Purchases Filtering
This commit revamps the Order Details screen, transforming it into a more interactive and informative page. It also refines the filtering logic on the Purchases screen and introduces navigation to the new "Refer a Friend" feature.

**Key Changes:**

*   **Revamped Order Details (`OrderDetailsFragment.kt`, `fragment_order_details.xml`):**
    *   **New Layout & Interactivity:** The layout has been significantly updated with better spacing, consistent styling, and interactive elements.
        *   Users can now directly message the seller, navigate to the seller's profile, get help, or view shipping status from this screen.
        *   Added a "View Product Details" expandable section to show more product information like category and price without leaving the screen.
    *   **Improved Data Display:**
        *   The order timestamp is now formatted to be more user-friendly (e.g., "Jan 15, 2025 at 09:30 AM").
        *   The overall UI has been polished with consistent margins and styling for a cleaner look.
    *   **New Navigation:**
        *   Added navigation from Order Details to `OrderStatusFragment`.
        *   Added navigation to the "Refer a Friend" screen (`TutorialsActivity`).

*   **Enhanced Purchases Filtering (`PurchasesFragment.kt`, `DashViewModel.kt`, `ApiInterface.kt`):**
    *   The `getProductsByStatus` API endpoint now accepts a `status` parameter to filter orders more effectively.
    *   The filter options on the Purchases screen have been updated to "All", "In Progress", and "Completed".
    *   The logic in `PurchasesFragment` was updated to use the new `status` parameter, allowing for more precise filtering of purchased items.

*   **UI and Code Refinements:**
    *   **`ChatActivity.kt`**: Commented out the logic that resets the unread message count upon opening a chat.
    *   **`TutorialsActivity.kt`**: Added logic to navigate to the `ReferFriendFragment` when the activity is opened with the type "refer".
    *   **`PurchasesAdapter.kt`**: Corrected the date format parsing for the "Purchased" date to handle the API's timestamp format.
    *   **`colors.xml`**: Adjusted the `onSurfaceVariant` color for UI consistency.
    *   **`ic_vehicle.xml`**: Resized the vehicle icon for better alignment in the UI.

+++ app/src/main/java/io/bidswipe/app/network/response/FetchBidResponse.kt
@@ -25,7 +25,7 @@ data class FetchBidResponse(
 	@Keep
 	data class Data(
 		@SerializedName("bid_price")
-		val bidPrice : Int? ,
+		val bidPrice : Double? ,
 		@SerializedName("created_at")
 		val createdAt : String? ,
 		@SerializedName("created_by")
